[Gold I] Title: 서로소 조합, Time: 24 ms, Memory: 16528 KB -BaekjoonHub

+++ 백준/Gold/33916. 서로소 조합/서로소 조합.cc
@@ -2,7 +2,7 @@
 #pragma GCC optimize("O3,unroll-loops")
 #pragma GCC target("avx,avx2,fma")
 using namespace std;
-using ll = long long;
+using ll = int;
 using pll = pair<ll,ll>;
 using ld = long double;
 using pld = pair<ld,ld>;
@@ -208,7 +208,7 @@ void preprocess() {
                 while (i/num > 0) {
                     factor[i][j] += i/num;
                     num *= prime[j];
-                }   
+                }
             }
             else break;
         }
